splitted suggest new point behaviour
